requirements.txt
aiohttp>=3.9.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
